+++ modules/client/examples/01-client/18-is-dao-update-valid.ts
@@ -21,7 +21,7 @@ const actions: DaoAction[] = [
 ];
 
 // check if a dap update proposal is valid
-const isValid = client.methods.isDaoUpdateValid({
+const isValid = client.methods.isDaoUpdateProposalValid({
   daoAddress: "0x1234567890123456789012345678901234567890",
   actions,
 });